requirements.txt
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
category-encoders==1.3.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
Flask==1.0.2
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
joblib==0.13.1
jsonschema==3.0.0b3
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lightgbm==2.2.2
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.2
patsy==0.5.1
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.8
pydotplus==2.0.2
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.9
python-dateutil==2.7.5
pytz==2018.9
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
scikit-learn==0.20.2
scikit-optimize==0.5.2
scipy==1.2.0
Send2Trash==1.5.0
six==1.12.0
sklearn-pandas==1.8.0
statsmodels==0.9.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
xgboost==0.81
